+++ extensions/supernotes/CHANGELOG.md
@@ -1,4 +1,4 @@
 # Supernotes Changelog
 
 ## [Initial Version] - 2022-06-30
-- Includes "create", "recents", and "search" commands
+- Includes "create" and "search" commands